refactor(loader): extract EP additions into ep-* modules for easier upstream rebasing
Move all Elastic Path custom code out of upstream files into dedicated modules:
- ep-s3-cache.ts: S3 singleton, tryGetS3CacheEntry, _testonly
- ep-loader-export-opts.ts: LOADER_CODEGEN_OPTS_DEFAULTS, makeExportOpts

Upstream files (s3-util.ts, gen-code-bundle.ts) now re-export EP symbols and
delegate to the ep-* modules, keeping their diffs minimal for future merges.
